# docs/swagger.yaml
basePath: /api
definitions:
...
  common.ErrorDetailElement:
    properties:
      field:
        type: string
      message:
        type: string
      param:
        type: string
      struct_field:
        type: string
      tag:
        type: string
      value: {}
    type: object
  common.HttpError:
    properties:
      code:
        type: integer
      detail:
        items:
          $ref: '#/definitions/common.ErrorDetailElement'
        type: array
      message:
        type: string
    type: object
...
  item.CreateItemModel:
    properties:
      commodity_id:
        type: integer
      item_name:
        type: string
      platform_id:
        type: integer
      price:
        type: number
      seller_id:
        type: integer
    required:
    - commodity_id
    - item_name
    - platform_id
    - price
    type: object
  item.SearchQuery:
    properties:
      accurate:
        type: boolean
      range:
        type: string
      search:
        type: string
    type: object
  item.UpdateItemModel:
    properties:
      commodity_id:
        type: integer
      commodity_item_id:
        type: integer
      item_name:
        type: string
      platform_id:
        type: integer
      price:
        type: number
      seller_id:
        type: integer
    required:
    - commodity_id
    - commodity_item_id
    - item_name
    - platform_id
    - price
    type: object
  models.Commodity:
    properties:
      category:
        type: string
      default_name:
        type: string
      id:
        type: integer
      produce_address:
        type: string
      produce_at:
        $ref: '#/definitions/models.MyTime'
    type: object
  models.CommodityItem:
    properties:
      commodity:
        $ref: '#/definitions/models.Commodity'
      commodity_id:
        type: integer
      id:
        type: integer
      item_name:
        type: string
      platform:
        $ref: '#/definitions/models.Platform'
      platform_id:
        type: integer
      price:
        type: number
      seller:
        $ref: '#/definitions/models.Seller'
      seller_id:
        type: integer
      update_at:
        $ref: '#/definitions/models.MyTime'
    type: object
...
  models.MyTime:
    properties:
      time.Time:
        type: string
    type: object
  models.Platform:
    properties:
      country:
        type: string
      id:
        type: integer
      name:
        type: string
      url:
        type: string
    type: object
...
  models.Seller:
    properties:
      address:
        description: Age      int
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
...
  platform.CreatePlatformRequest:
    properties:
      country:
        type: string
      name:
        type: string
      url:
        type: string
    type: object
...
info:
  contact:
    email: polaris@fduhole.com
    name: Maintainer Polaris
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Database System Design Project
  version: 1.0.0
paths:
...
  /api/commodity/all:
    get:
      consumes:
      - application/json
      description: 获取所有商品
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CommodityItem'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 获取所有商品
      tags:
      - Item
  /api/commodity/data:
    get:
      consumes:
      - application/json
      description: 获取自己的商品
      parameters:
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CommodityItem'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 获取自己的商品
      tags:
      - Item
  /api/commodity/item:
    delete:
      consumes:
      - application/json
      description: 删除商品
      parameters:
      - description: item id
        in: path
        name: id
        required: true
        type: string
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 删除商品
      tags:
      - Item
    post:
      consumes:
      - application/json
      description: 添加商品
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/item.CreateItemModel'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 添加商品
      tags:
      - Item
    put:
      consumes:
      - application/json
      description: 更新商品
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/item.UpdateItemModel'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 更新商品
      tags:
      - Item
  /api/commodity/item/{id}:
    get:
      consumes:
      - application/json
      description: 根据商品ID获取多个商品项
      parameters:
      - description: commodity_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CommodityItem'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 根据商品ID获取多个商品项
      tags:
      - Item
  /api/commodity/item/batch:
    post:
      consumes:
      - application/json
      description: 批量添加商品
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          items:
            $ref: '#/definitions/item.CreateItemModel'
          type: array
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: 批量添加商品
      tags:
      - Item
...
  /api/platform:
    delete:
      consumes:
      - application/json
      description: Delete platform
      parameters:
      - description: platform id
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Delete platform
      tags:
      - Platform
    post:
      consumes:
      - application/json
      description: Add platform
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/platform.CreatePlatformRequest'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Add platform
      tags:
      - Platform
    put:
      consumes:
      - application/json
      description: Update platform
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.Platform'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Update platform
      tags:
      - Platform
...
  /api/search:
    post:
      consumes:
      - application/json
      description: 搜索商品
      parameters:
      - description: 'Range: name, category; Search: content to search'
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/item.SearchQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CommodityItem'
            type: array
      summary: 搜索商品
      tags:
      - Item
...
swagger: "2.0"